ig/local-gadget: Rename local-gadget in the tests
