+++ Dockerfiles/local-gadget-tests.Dockerfile
@@ -1,5 +1,5 @@
 FROM ghcr.io/inspektor-gadget/ig-builder:latest
 
-WORKDIR /go/src/github.com/inspektor-gadget/inspektor-gadget/integration/local-gadget/k8s
+WORKDIR /go/src/github.com/inspektor-gadget/inspektor-gadget/integration/ig/k8s
 
 RUN go test -c -o local-gadget-integration.test ./...
