+++ integration/ig/k8s/top_tcp_test.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Inspektor Gadget authors
+// Copyright 2022-2023 The Inspektor Gadget authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,20 +16,21 @@ package main
 
 import (
 	"fmt"
+	"syscall"
 	"testing"
 
 	. "github.com/inspektor-gadget/inspektor-gadget/integration"
-	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/top/file/types"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/top/tcp/types"
 	eventtypes "github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
 
-func TestTopFile(t *testing.T) {
+func TestTopTCP(t *testing.T) {
 	t.Parallel()
-	ns := GenerateTestNamespaceName("test-top-file")
+	ns := GenerateTestNamespaceName("test-top-tcp")
 
-	topFileCmd := &Command{
-		Name:         "TopFile",
-		Cmd:          fmt.Sprintf("ig top file -o json -m 999 --runtimes=%s", *containerRuntime),
+	topTCPCmd := &Command{
+		Name:         "TopTCP",
+		Cmd:          fmt.Sprintf("ig top tcp -o json -m 999 --runtimes=%s", *containerRuntime),
 		StartAndStop: true,
 		ExpectedOutputFn: func(output string) error {
 			expectedEntry := &types.Stats{
@@ -37,21 +38,20 @@ func TestTopFile(t *testing.T) {
 					Namespace: ns,
 					Pod:       "test-pod",
 				},
-				// echo is built-in
-				Comm:     "sh",
-				Filename: "bar",
-				FileType: 'R',
+				Comm:   "curl",
+				Family: syscall.AF_INET,
+				Dport:  80,
+				Saddr:  "127.0.0.1",
+				Daddr:  "127.0.0.1",
 			}
 
 			normalize := func(e *types.Stats) {
 				e.Container = ""
 				e.Pid = 0
-				e.Tid = 0
 				e.MountNsID = 0
-				e.Reads = 0
-				e.ReadBytes = 0
-				e.Writes = 0
-				e.WriteBytes = 0
+				e.Sport = 0
+				e.Sent = 0
+				e.Received = 0
 			}
 
 			return ExpectEntriesInMultipleArrayToMatch(output, normalize, expectedEntry)
@@ -60,9 +60,9 @@ func TestTopFile(t *testing.T) {
 
 	commands := []*Command{
 		CreateTestNamespaceCommand(ns),
-		topFileCmd,
-		SleepForSecondsCommand(2), // wait to ensure local-gadget has started
-		BusyboxPodRepeatCommand(ns, "echo foo > bar"),
+		topTCPCmd,
+		SleepForSecondsCommand(2), // wait to ensure ig has started
+		PodCommand("test-pod", "nginx", ns, "[sh, -c]", "nginx && while true; do curl 127.0.0.1; sleep 0.1; done"),
 		WaitUntilTestPodReadyCommand(ns),
 		DeleteTestNamespaceCommand(ns),
 	}
