+++ integration/ig/k8s/trace_capabilities_test.go
@@ -19,42 +19,33 @@ import (
 	"testing"
 
 	. "github.com/inspektor-gadget/inspektor-gadget/integration"
-	execTypes "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/exec/types"
+	capabilitiesTypes "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/trace/capabilities/types"
 )
 
-func TestTraceExec(t *testing.T) {
+func TestTraceCapabilities(t *testing.T) {
 	t.Parallel()
-	ns := GenerateTestNamespaceName("test-trace-exec")
+	ns := GenerateTestNamespaceName("test-trace-capabilities")
 
-	cmd := "while true; do date ; /bin/sleep 0.1; done"
-	shArgs := []string{"/bin/sh", "-c", cmd}
-	dateArgs := []string{"/bin/date"}
-	sleepArgs := []string{"/bin/sleep", "0.1"}
-
-	traceExecCmd := &Command{
-		Name:         "TraceExec",
-		Cmd:          fmt.Sprintf("ig trace exec -o json --runtimes=%s", *containerRuntime),
+	capabilitiesCmd := &Command{
+		Name:         "TraceCapabilities",
+		Cmd:          fmt.Sprintf("ig trace capabilities -o json --runtimes=%s", *containerRuntime),
 		StartAndStop: true,
 		ExpectedOutputFn: func(output string) error {
-			expectedEntries := []*execTypes.Event{
-				{
-					Event: BuildBaseEvent(ns),
-					Comm:  "sh",
-					Args:  shArgs,
-				},
-				{
-					Event: BuildBaseEvent(ns),
-					Comm:  "date",
-					Args:  dateArgs,
-				},
-				{
-					Event: BuildBaseEvent(ns),
-					Comm:  "sleep",
-					Args:  sleepArgs,
-				},
+			expectedEntry := &capabilitiesTypes.Event{
+				Event:         BuildBaseEvent(ns),
+				Comm:          "nice",
+				CapName:       "SYS_NICE",
+				Cap:           23,
+				Syscall:       "setpriority",
+				Audit:         1,
+				Verdict:       "Deny",
+				CurrentUserNs: 1,
+				TargetUserNs:  1,
+				Caps:          1,
+				CapsNames:     []string{"x"},
 			}
 
-			normalize := func(e *execTypes.Event) {
+			normalize := func(e *capabilitiesTypes.Event) {
 				// TODO: Handle it once we support getting K8s container name for docker
 				// Issue: https://github.com/inspektor-gadget/inspektor-gadget/issues/737
 				if *containerRuntime == ContainerRuntimeDocker {
@@ -63,21 +54,34 @@ func TestTraceExec(t *testing.T) {
 
 				e.Timestamp = 0
 				e.Pid = 0
-				e.Ppid = 0
 				e.UID = 0
-				e.Retval = 0
 				e.MountNsID = 0
+				// Do not check InsetID to avoid introducing dependency on the kernel version
+				e.InsetID = nil
+
+				if e.CurrentUserNs != 0 {
+					e.CurrentUserNs = 1
+				}
+				if e.TargetUserNs != 0 {
+					e.TargetUserNs = 1
+				}
+				if e.Caps > 0 {
+					e.Caps = 1
+				}
+				if len(e.CapsNames) != 0 {
+					e.CapsNames = []string{"x"}
+				}
 			}
 
-			return ExpectEntriesToMatch(output, normalize, expectedEntries...)
+			return ExpectEntriesToMatch(output, normalize, expectedEntry)
 		},
 	}
 
 	commands := []*Command{
 		CreateTestNamespaceCommand(ns),
-		traceExecCmd,
-		SleepForSecondsCommand(2), // wait to ensure local-gadget has started
-		BusyboxPodCommand(ns, cmd),
+		capabilitiesCmd,
+		SleepForSecondsCommand(2), // wait to ensure ig has started
+		BusyboxPodRepeatCommand(ns, "nice -n -20 echo"),
 		WaitUntilTestPodReadyCommand(ns),
 		DeleteTestNamespaceCommand(ns),
 	}
