+++ integration/ig/k8s/trace_mount_test.go
@@ -65,7 +65,7 @@ func TestTraceMount(t *testing.T) {
 	commands := []*Command{
 		CreateTestNamespaceCommand(ns),
 		traceMountCmd,
-		SleepForSecondsCommand(2), // wait to ensure local-gadget has started
+		SleepForSecondsCommand(2), // wait to ensure ig has started
 		BusyboxPodRepeatCommand(ns, "mount /mnt /mnt"),
 		WaitUntilTestPodReadyCommand(ns),
 		DeleteTestNamespaceCommand(ns),
