+++ integration/ig/k8s/trace_oomkill_test.go
@@ -79,7 +79,7 @@ spec:
 	commands := []*Command{
 		CreateTestNamespaceCommand(ns),
 		traceOOMKillCmd,
-		SleepForSecondsCommand(2), // wait to ensure local-gadget has started
+		SleepForSecondsCommand(2), // wait to ensure ig has started
 		{
 			Name:           "RunOomkillTestPod",
 			Cmd:            fmt.Sprintf("echo '%s' | kubectl apply -f -", limitPodYaml),
