+++ integration/ig/non-k8s/trace_dns_test.go
